+++ src/soundwrapper.c
@@ -441,7 +441,7 @@ void sub_80029E8(int songId, int b, int c, int d, ProcPtr parent)
     mproc->unk66 = c;
 }
 
-int MusicProc4Exists(void)
+s8 MusicProc4Exists(void)
 {
     if (Proc_Find(sMusicProc4Script) != NULL)
         return TRUE;
